requirements.txt
openai==1.12.0
python-dotenv==1.0.0
numpy>=2.3.3
colorama==0.4.6
tabulate==0.9.0
tenacity==8.2.3 
chromadb==1.0.15